# turning json option into macro parameter