--- a/benchmarks/benchmark.cpp
+++ b/benchmarks/benchmark.cpp
@@ -236,7 +236,7 @@ int main(int argc, char **argv) {
               << std::endl;
 #endif
   }
-  if(argc > 1) {
+  if (argc > 1) {
     fileload(argv[1]);
     return EXIT_SUCCESS;
   }