@dalle 2025-03-09
Looks good. Do you think there be a macro for if-constexpr for C++17 enabled compilers?

Something like this?
```c++
#if __cplusplus >= 201703L || (defined(_MSVC_LANG) && _MSVC_LANG >= 201703L)
#define FASTFLOAT_IF_CONSTEXPR17(x) if constexpr (x)
#else
#define FASTFLOAT_IF_CONSTEXPR17(x) if (x)
#endif

...
FASTFLOAT_IF_CONSTEXPR17(basic_json_fmt)
```