@IRainman 2025-03-09
Nice! Also the code definitely more cache friendly, it's shown by more constant time results on output.