@IRainman 2025-03-09
Also about std::from_chars and compatibility, as you can remember, the standard regexp isn't ideal and the fmt library, especially with compile time parsing macros 
`fmt::format(FMT_COMPILE("some value: {}"), value));`
much faster than std implementation and generates very compact code.