@IRainman 2025-03-09
> Looks good. I commented somewhere else that we could explore using if-constexpr, but it's up to you. I'm a bit worried of "constant expression in if-statement" warnings appearing.

It's probably fine because it's old code and in the std function should be consteval in used in constexpr if. Maybe I should write a defect report for standard std::is_constant_evaluated. What do you think?