@IRainman 2025-03-09
Maybe it's better to convert all options into a template parameter?