@lemire 2025-03-10
> Maybe it's better to convert all options into a template parameter?

Well. That's an empirical question (i.e., see benchmarks).